+++ shoop/core/order_creator/_source.py
@@ -7,6 +7,7 @@
 # LICENSE file in the root directory of this source tree.
 from __future__ import unicode_literals
 
+from django.utils.encoding import force_text
 from django.utils.timezone import now
 
 from shoop.core import taxing
@@ -19,6 +20,7 @@ from shoop.core.taxing import TaxableItem
 from shoop.utils.decorators import non_reentrant
 from shoop.utils.money import Money
 
+from ._source_modifier import get_order_source_modifier_modules
 from .signals import post_compute_source_lines
 
 
@@ -116,6 +118,7 @@ class OrderSource(object):
         self.shipping_data = {}
         self.extra_data = {}
 
+        self._codes = []
         self._lines = []
 
         self.zero_price = shop.create_price(0)
@@ -178,7 +181,7 @@ class OrderSource(object):
     taxful_total_discount_or_none = taxful_total_discount.or_none
     taxless_total_discount_or_none = taxless_total_discount.or_none
 
-    total_price_of_products = _PriceSum("price", "_get_product_lines")
+    total_price_of_products = _PriceSum("price", "get_product_lines")
 
     @property
     def shipping_method(self):
@@ -207,6 +210,61 @@ class OrderSource(object):
     def status(self, status):
         self.status_id = (status.id if status else None)
 
+    @property
+    def codes(self):
+        return list(self._codes)
+
+    def add_code(self, code):
+        """
+        Add code to this OrderSource.
+
+        At this point it is expected that the customers
+        permission to use the code has already been
+        checked by the caller.
+
+        The code will be converted to text.
+
+        :param code: The code to add
+        :type code: str
+        :return: True if code was added, False if it was already there
+        :rtype: bool
+        """
+        code_text = force_text(code)
+        if code_text not in self._codes:
+            self._codes.append(code_text)
+            self.uncache()
+            return True
+        return False
+
+    def clear_codes(self):
+        """
+        Remove all codes from this OrderSource.
+
+        :return: True iff there was codes before clearing
+        :rtype: bool
+        """
+        if self._codes:
+            self._codes = []
+            self.uncache()
+            return True
+        return False
+
+    def remove_code(self, code):
+        """
+        Remove given code from this OrderSource.
+
+        :param code: The code to remove
+        :type code: str
+        :return: True if code was removed, False if code was not there
+        :rtype: bool
+        """
+        code_text = force_text(code)
+        if code_text in self._codes:
+            self._codes.remove(code_text)
+            self.uncache()
+            return True
+        return False
+
     def add_line(self, **kwargs):
         line = SourceLine(source=self, **kwargs)
         self._lines.append(line)
@@ -220,6 +278,15 @@ class OrderSource(object):
         See also `get_final_lines`.
         """
         return self._lines
+
+    @property
+    def product_count(self):
+        """
+        Get the total number of products in this OrderSource.
+
+        :rtype: decimal.Decimal|int
+        """
+        return sum([line.quantity for line in self.get_product_lines()])
 
     def get_final_lines(self, with_taxes=False):
         """
@@ -284,6 +351,7 @@ class OrderSource(object):
 
         lines.extend(self._compute_payment_method_lines())
         lines.extend(self._compute_shipping_method_lines())
+        self._add_lines_from_modifiers(lines)
 
         lines.extend(_collect_lines_from_signal(
             post_compute_source_lines.send(
@@ -301,7 +369,16 @@ class OrderSource(object):
             for line in self.shipping_method.get_source_lines(self):
                 yield line
 
-    def _get_product_lines(self):
+    def _add_lines_from_modifiers(self, lines):
+        """
+        Add lines from OrderSourceModifiers to given list of lines.
+        """
+        for module in get_order_source_modifier_modules():
+            new_lines = list(module.get_new_lines(self, list(lines)))
+            # Extend lines now to allow next module to see them
+            lines.extend(new_lines)
+
+    def get_product_lines(self):
         """
         Get lines with a product.