+++ shoop_tests/core/test_default_pricing.py
@@ -14,6 +14,7 @@ from shoop.testing.factories import (
     create_product, create_random_person, get_default_customer_group,
     get_default_shop
 )
+from shoop.testing.utils import apply_request_middleware
 
 original_pricing_module = settings.SHOOP_PRICING_MODULE
 
@@ -43,6 +44,7 @@ def initialize_test(rf, include_tax=False):
 
     request = rf.get("/")
     request.shop = shop
+    apply_request_middleware(request)
     request.customer = customer
     return request, shop, group