+++ shoop_tests/core/test_product_variations.py
@@ -9,10 +9,9 @@
 import pytest
 
 from shoop.core.models import (
-    ProductVariationResult, ProductVariationVariable,
+    ProductMode, ProductVariationResult, ProductVariationVariable,
     ProductVariationVariableValue, ShopProduct
 )
-from shoop.core.models import ProductMode
 from shoop.testing.factories import create_product, get_default_shop